+++ src/libclient/timing/timing.cpp
@@ -1,6 +1,6 @@
 #include "timing.h"
 
-int Timing::timeLeft(const QDateTime &when) const
+qint64 Timing::timeLeft(const QDateTime &when) const
 {
     return when.msecsTo(nextRun());
 }
